git: unpack-objects
