--- completers/git_completer/cmd/unpack_objects_generated.go
@@ -1,19 +1,0 @@
-package cmd
-
-import (
-	"github.com/carapace-sh/carapace"
-	"github.com/spf13/cobra"
-)
-
-var unpack_objectsCmd = &cobra.Command{
-	Use:     "unpack-objects",
-	Short:   "Unpack objects from a packed archive",
-	Run:     func(cmd *cobra.Command, args []string) {},
-	GroupID: groups[group_low_level_manipulator].ID,
-}
-
-func init() {
-	carapace.Gen(unpack_objectsCmd).Standalone()
-
-	rootCmd.AddCommand(unpack_objectsCmd)
-}
